fix(optimum): drop the redundant deepstack cast in qwen3_vl
Same pattern this branch already removed from qwen2_vl: the cached
deepstack embeds are visual-runtime outputs, so they are already in the
compiled dtype. qwen3_vl.py landed on dev while #800 was reverted and
carried the cast over.

+++ vllm_rbln/model_executor/models/optimum/qwen3_vl.py
@@ -277,7 +277,7 @@ class RBLNOptimumQwen3VLForConditionalGeneration(
                 return None
             num_layers = len(present[0][key])
             return [
-                torch.cat([c[key][layer].to(self.dtype) for c in present], dim=0)
+                torch.cat([c[key][layer] for c in present], dim=0)
                 for layer in range(num_layers)
             ]
